Add file via upload
These results in a return of 2 new lists of each year: one of stemmed tokens, and another of lemmatiz.
All text is put on a level playing field after converting all text to the same case (upper or lower), removing punctuation, converting numbers to their word equivalents, and so on. Normalization puts all words on equal footing, and allows processing to proceed uniformly.